refuc(crypto-helper): asn1: replace some strings components with macro generation;
