+++ src/asn1/scheme/strings.rs
@@ -9,26 +9,6 @@ use crate::asn1::scheme::build_asn1_schema;
 use crate::asn1::HighlightAction;
 
 #[derive(PartialEq, Properties, Clone)]
-pub struct Utf8StringNodeProps {
-    pub node: OwnedUtf8String,
-    pub meta: OwnedRawAsn1EntityData,
-}
-
-#[function_component(Utf8StringNode)]
-pub fn utf8_string(props: &Utf8StringNodeProps) -> Html {
-    let offset = props.meta.tag_position();
-    let length_len = props.meta.length_range().len();
-    let data_len = props.meta.data_range().len();
-
-    html! {
-        <div class="terminal-asn1-node">
-            <NodeOptions node_bytes={props.meta.raw_bytes().to_vec()} {offset} {length_len} {data_len} name={String::from("UTF8String")}/>
-            <span class="asn-simple-value">{props.node.string().to_owned()}</span>
-        </div>
-    }
-}
-
-#[derive(PartialEq, Properties, Clone)]
 pub struct OctetStringNodeProps {
     pub node: OwnedOctetString,
     pub meta: OwnedRawAsn1EntityData,
@@ -48,7 +28,7 @@ pub fn octet_string(props: &OctetStringNodeProps) -> Html {
         Some(asn1) => html! {
             <div style="cursor: crosshair">
                 <div class="asn1-constructor-header">
-                    <NodeOptions node_bytes={props.meta.raw_bytes().to_vec()} {offset} {length_len} {data_len} name={String::from("Octet String")}/>
+                    <NodeOptions node_bytes={props.meta.raw_bytes().to_vec()} {offset} {length_len} {data_len} name={String::from("OctetString")}/>
                 </div>
                 <div class="asn1-constructor-body">
                     {build_asn1_schema(asn1, &props.cur_node, &props.set_cur_node)}
@@ -62,7 +42,7 @@ pub fn octet_string(props: &OctetStringNodeProps) -> Html {
             };
             html! {
                 <div class="terminal-asn1-node">
-                    <NodeOptions node_bytes={props.meta.raw_bytes().to_vec()} {offset} {length_len} {data_len} name={String::from("Octet String")} />
+                    <NodeOptions node_bytes={props.meta.raw_bytes().to_vec()} {offset} {length_len} {data_len} name={String::from("OctetString")} />
                     <span class="asn1-node-info-label">{format!("({} bytes)", octets.len())}</span>
                     <span class="asn-simple-value">{encoded_octets}</span>
                 </div>
@@ -94,7 +74,7 @@ pub fn bit_string(props: &BitStringNodeProps) -> Html {
 
     html! {
         <div class="terminal-asn1-node">
-            <NodeOptions node_bytes={props.meta.raw_bytes().to_vec()} {offset} {length_len} {data_len} name={String::from("Bit String")} />
+            <NodeOptions node_bytes={props.meta.raw_bytes().to_vec()} {offset} {length_len} {data_len} name={String::from("BitString")} />
             <span class="asn1-node-info-label">{format!("({} bits)", bits_amount)}</span>
             <span class="asn-simple-value">{bits}</span>
         </div>
@@ -123,68 +103,13 @@ pub fn bit_string(props: &BmpStringNodeProps) -> Html {
 
     html! {
         <div class="terminal-asn1-node">
-            <NodeOptions node_bytes={props.meta.raw_bytes().to_vec()} {offset} {length_len} {data_len} name={String::from("Bmp String")} />
+            <NodeOptions node_bytes={props.meta.raw_bytes().to_vec()} {offset} {length_len} {data_len} name={String::from("BmpString")} />
             <span class="asn-simple-value">{s}</span>
         </div>
     }
 }
 
-#[derive(PartialEq, Properties, Clone)]
-pub struct IA5StringNodeProps {
-    pub node: OwnedIA5String,
-    pub meta: OwnedRawAsn1EntityData,
-}
-
-#[function_component(IA5StringNode)]
-pub fn utf8_string(props: &IA5StringNodeProps) -> Html {
-    let offset = props.meta.tag_position();
-    let length_len = props.meta.length_range().len();
-    let data_len = props.meta.data_range().len();
-
-    html! {
-        <div class="terminal-asn1-node">
-            <NodeOptions node_bytes={props.meta.raw_bytes().to_vec()} {offset} {length_len} {data_len} name={String::from("IA5String")}/>
-            <span class="asn-simple-value">{props.node.string().to_owned()}</span>
-        </div>
-    }
-}
-
-#[derive(PartialEq, Properties, Clone)]
-pub struct PrintableStringNodeProps {
-    pub node: OwnedPrintableString,
-    pub meta: OwnedRawAsn1EntityData,
-}
-
-#[function_component(PrintableStringNode)]
-pub fn utf8_string(props: &PrintableStringNodeProps) -> Html {
-    let offset = props.meta.tag_position();
-    let length_len = props.meta.length_range().len();
-    let data_len = props.meta.data_range().len();
-
-    html! {
-        <div class="terminal-asn1-node">
-            <NodeOptions node_bytes={props.meta.raw_bytes().to_vec()} {offset} {length_len} {data_len} name={String::from("PrintableString")}/>
-            <span class="asn-simple-value">{props.node.string().to_owned()}</span>
-        </div>
-    }
-}
-
-#[derive(PartialEq, Properties, Clone)]
-pub struct GeneralStringNodeProps {
-    pub node: OwnedGeneralString,
-    pub meta: OwnedRawAsn1EntityData,
-}
-
-#[function_component(GeneralStringNode)]
-pub fn utf8_string(props: &GeneralStringNodeProps) -> Html {
-    let offset = props.meta.tag_position();
-    let length_len = props.meta.length_range().len();
-    let data_len = props.meta.data_range().len();
-
-    html! {
-        <div class="terminal-asn1-node">
-            <NodeOptions node_bytes={props.meta.raw_bytes().to_vec()} {offset} {length_len} {data_len} name={String::from("GeneralString")}/>
-            <span class="asn-simple-value">{props.node.string().to_owned()}</span>
-        </div>
-    }
-}
+define_string_node!(GeneralString);
+define_string_node!(IA5String);
+define_string_node!(PrintableString);
+define_string_node!(Utf8String);
